+++ src/components/containers/ProfileListBoxContainer.jsx
@@ -1,7 +1,9 @@
 import { useEffect, useState } from 'react';
 
-import { getAnswerers } from '@services/api/api';
-import CardList from '@ui/CardList';
+import ToggleMenuContainer from './ToggleMenuContainer';
+import { getAnswerers } from '@services/api/get.js';
+import PaginationBar from '@ui/PaginationBar';
+import ProfileList from '@ui/ProfileList';
 import { isTabletMini } from '@utils/windowSize';
 
 function ProfileListBoxContainer() {
@@ -48,9 +50,9 @@ function ProfileListBoxContainer() {
 
   return (
     <div>
-      {/* 정렬버튼에 prop으로 order state내려줍니다 */}
-      <CardList profileList={profileList} />
-      {/* 페이지네이션 */}
+      <ToggleMenuContainer onClick={handleOrder} />
+      <ProfileList profileList={profileList} />
+      <PaginationBar />
     </div>
   );
 }
