+++ src/components/css/PaginationBar.module.css
@@ -1,1 +1,47 @@
+.paginationBar {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-top: 40px;
+}
 
+.paginationButton[disabled] {
+  cursor: none;
+}
+
+.paginationButton,
+.currentPageButton {
+  border: none;
+  width: 40px;
+  height: 40px;
+  color: var(--color-gray40);
+  font-weight: 400;
+  font-size: 20px;
+  line-height: 25px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background-color: rgba(0, 0, 0, 0);
+}
+
+.currentPageButton {
+  color: var(--color-brown40);
+}
+
+@media (max-width: 1199px) {
+  .paginationBar {
+    margin-top: 46px;
+  }
+}
+
+@media (min-width: 768px) and (max-width: 883px) {
+  .paginationBar {
+    margin-top: 61px;
+  }
+}
+
+@media (max-width: 767px) {
+  .paginationBar {
+    margin-top: 31px;
+  }
+}
