+++ src/components/ui/Button.jsx
@@ -1,20 +1,17 @@
 import arrowRightIcon from '../../assets/images/ic_arrow_right.svg';
 import styles from '../css/Button.module.css';
 
-
-const Button = ({ onClick, children, variant, size }) => {
-// const Button = ({ onClick, isDisabled, children, variant, type }) => {
+const Button = ({ onClick, isDisabled, children, variant, type, size }) => {
   const buttonClassName = `${styles.button} ${styles[variant] || ''}`;
   const sizeClassName = `${styles[size] || ''}`;
 
   return (
-    <button className={`${buttonClassName} ${sizeClassName}`} onClick={onClick}>
-//     <button
-//       type={type}
-//       disabled={isDisabled}
-//       className={buttonClassName}
-//       onClick={onClick}
-//     >
+    <button
+      type={type}
+      disabled={isDisabled}
+      className={`${buttonClassName} ${sizeClassName}`}
+      onClick={onClick}
+    >
       {children}
       {variant === 'outline' && (
         <img src={arrowRightIcon} alt="화살표 아이콘" />
