+++ src/components/ui/PaginationBar.jsx
@@ -1,9 +1,27 @@
-import LeftButton from '../../assets/images/ic_left_button.svg';
-import RightButton from '../../assets/images/ic_right_button.svg';
-import styles from '../../components/css/PaginationBar.module.css';
+import styles from '../css/PaginationBar.module.css';
 
 const PaginationBar = () => {
-  return;
+  const pages = [1, 2, 3, 4, 5];
+  const currentPage = 1;
+
+  return (
+    <div className={styles.paginationBar}>
+      <button className={styles.paginationButton}>{'<'}</button>
+      {pages.map((page) => (
+        <button
+          key={page}
+          className={
+            currentPage === page
+              ? styles.currentPageButton
+              : styles.paginationButton
+          }
+        >
+          {page}
+        </button>
+      ))}
+      <button className={styles.paginationButton}>{'>'}</button>
+    </div>
+  );
 };
 
 export default PaginationBar;
