+++ src/pages/ListPage.jsx
@@ -1,9 +1,7 @@
 import { Link } from 'react-router-dom';
 
-import PaginationBar from '@ui/PaginationBar';
-import ToggleMenu from '@ui/ToggleMenu';
-import ProfileListBoxContainer from '@containers/ProfileListBoxContainer';
 import styles from './css/ListPage.module.css';
+import ProfileListBoxContainer from '@containers/ProfileListBoxContainer';
 import Button from '@ui/Button';
 // import CardList from '@ui/CardList';
 import Logo from '@ui/Logo';
@@ -19,10 +17,8 @@ function ListPage() {
           <Button variant={'outline'}>답변하러 가기</Button>
         </div>
         <h2 className={styles.title}>누구에게 질문할까요?</h2>
-        <ToggleMenu />
         <ProfileListBoxContainer />
       </div>
-      <PaginationBar></PaginationBar>
     </main>
   );
 }
